@@ -4,7 +4,7 @@
 # See MIT-LICENSE file or http://www.opensource.org/licenses/mit-license.php
 #------------------------------------------------------------------------------
 class Admin::UsersController < Admin::ApplicationController
-  before_action "set_current_tab('admin/users')", only: [:index, :show]
+  before_action :setup_current_tab, only: [:index, :show]
 
   load_resource except: [:create]
 
@@ -44,8 +44,6 @@ class Admin::UsersController < Admin::ApplicationController
   # POST /admin/users.xml                                                  AJAX
   #----------------------------------------------------------------------------
   def create
-    params[:user][:email].try(:strip!)
-    params[:user][:password_confirmation] = nil if params[:user][:password_confirmation].blank?
     @user = User.new(user_params)
     @user.check_if_needs_approval
     @user.save_without_session_maintenance
@@ -57,8 +55,6 @@ class Admin::UsersController < Admin::ApplicationController
   # PUT /admin/users/1.xml                                                 AJAX
   #----------------------------------------------------------------------------
   def update
-    params[:user][:email].try(:strip!)
-    params[:user][:password_confirmation] = nil if params[:user][:password_confirmation].blank?
     @user = User.find(params[:id])
     @user.attributes = user_params
     @user.save_without_session_maintenance
@@ -108,6 +104,10 @@ class Admin::UsersController < Admin::ApplicationController
   protected
 
   def user_params
+    return {} unless params[:user]
+    params[:user][:email].try(:strip!)
+    params[:user][:password_confirmation] = nil if params[:user][:password_confirmation].blank?
+
     params[:user].permit(
       :admin,
       :username,
@@ -136,7 +136,7 @@ class Admin::UsersController < Admin::ApplicationController
     self.current_page  = options[:page] if options[:page]
     self.current_query = params[:query] if params[:query]
 
-    @search = klass.search(params[:q])
+    @search = klass.ransack(params[:q])
     @search.build_grouping unless @search.groupings.any?
 
     wants = request.format
@@ -145,5 +145,9 @@ class Admin::UsersController < Admin::ApplicationController
     scope = scope.text_search(current_query)      if current_query.present?
     scope = scope.paginate(page: current_page) if wants.html? || wants.js? || wants.xml?
     scope
+  end
+
+  def setup_current_tab
+    set_current_tab('admin/users')
   end
 end
