@@ -45,7 +45,7 @@ class Campaign < ActiveRecord::Base
   scope :created_by,  ->(user) { where(user_id: user.id) }
   scope :assigned_to, ->(user) { where(assigned_to: user.id) }
 
-  scope :text_search, ->(query) { search('name_cont' => query).result }
+  scope :text_search, ->(query) { ransack('name_cont' => query).result }
 
   uses_user_permissions
   acts_as_commentable
