@@ -58,7 +58,7 @@ class Lead < ActiveRecord::Base
   scope :created_by,   ->(user) { where(user_id: user.id) }
   scope :assigned_to,  ->(user) { where(assigned_to: user.id) }
 
-  scope :text_search, ->(query) { search('first_name_or_last_name_or_company_or_email_cont' => query).result }
+  scope :text_search, ->(query) { ransack('first_name_or_last_name_or_company_or_email_cont' => query).result }
 
   uses_user_permissions
   acts_as_commentable
