@@ -17,7 +17,7 @@ if defined?(FatFreeCRM::Application)
     config.action_controller.perform_caching = true
 
     # Disable Rails's static asset server (Apache or nginx will already do this)
-    config.serve_static_files = true
+    config.public_file_server.enabled = true
 
     # Compress JavaScripts and CSS
     config.assets.compress = true
