@@ -17,8 +17,8 @@ if defined?(FatFreeCRM::Application)
     config.eager_load = false
 
     # Configure static file server for tests with Cache-Control for performance.
-    config.serve_static_files   = true
-    config.static_cache_control = 'public, max-age=3600'
+    config.public_file_server.enabled = true
+    config.public_file_server.headers = { 'Cache-Control' => 'public, max-age=3600' }
 
     # Show full error reports and disable caching
     config.consider_all_requests_local       = true
