@@ -47,7 +47,7 @@ describe AuthenticationsController do
 
       describe "POST authentication" do
         it "displays 'must be logged out message' and redirects to profile page" do
-          post :create, authentication: @login
+          post :create, params: { authentication: @login }
           expect(flash[:notice]).not_to eq(nil)
           expect(flash[:notice]).to match(/^You must be logged out/)
           expect(response).to redirect_to(profile_path)
@@ -75,7 +75,7 @@ describe AuthenticationsController do
         @user = FactoryGirl.create(:user, username: "user", password: "pass", password_confirmation: "pass", login_count: 0)
         allow(@authentication).to receive(:user).and_return(@user)
 
-        post :create, authentication: @login
+        post :create, params: { authentication: @login }
         expect(flash[:notice]).not_to eq(nil)
         expect(flash[:notice]).not_to match(/last login/)
         expect(response).to redirect_to(root_path)
@@ -85,7 +85,7 @@ describe AuthenticationsController do
         @user = FactoryGirl.create(:user, username: "user", password: "pass", password_confirmation: "pass", login_count: 42)
         allow(@authentication).to receive(:user).and_return(@user)
 
-        post :create, authentication: @login
+        post :create, params: { authentication: @login }
         expect(flash[:notice]).to match(/last login/)
         expect(response).to redirect_to(root_path)
       end
@@ -99,7 +99,7 @@ describe AuthenticationsController do
           allow(@authentication).to receive(:save).and_return(false) # <--- Authentication failure.
           allow(Authentication).to receive(:new).and_return(@authentication)
 
-          post :create, authentication: @login
+          post :create, params: { authentication: @login }
           expect(flash[:warning]).not_to eq(nil)
           expect(response).to redirect_to(action: :new)
         end
@@ -116,7 +116,7 @@ describe AuthenticationsController do
           @user = FactoryGirl.create(:user, username: "user", password: "pass", password_confirmation: "pass", suspended_at: Date.yesterday, login_count: 0, last_login_at: nil, last_login_ip: nil)
           allow(@authentication).to receive(:user).and_return(@user)
 
-          post :create, authentication: @login
+          post :create, params: { authentication: @login }
           expect(@authentication.user.login_count).to eq(0)
           expect(@authentication.user.last_login_at).to be_nil
           expect(@authentication.user.last_login_ip).to be_nil
@@ -126,7 +126,7 @@ describe AuthenticationsController do
           @user = FactoryGirl.create(:user, username: "user", password: "pass", password_confirmation: "pass", suspended_at: Date.yesterday)
           allow(@authentication).to receive(:user).and_return(@user)
 
-          post :create, authentication: @login
+          post :create, params: { authentication: @login }
           expect(flash[:warning]).not_to eq(nil) # Invalid username/password.
           expect(flash[:notice]).to eq(nil)      # Not approved yet.
           expect(response).to redirect_to(action: :new)
@@ -137,7 +137,7 @@ describe AuthenticationsController do
           @user = FactoryGirl.create(:user, username: "user", password: "pass", password_confirmation: "pass", suspended_at: Date.yesterday, login_count: 0)
           allow(@authentication).to receive(:user).and_return(@user)
 
-          post :create, authentication: @login
+          post :create, params: { authentication: @login }
           expect(flash[:warning]).to eq(nil)     # Invalid username/password.
           expect(flash[:notice]).not_to eq(nil)  # Not approved yet.
           expect(response).to redirect_to(action: :new)
