@@ -55,7 +55,7 @@ describe AccountsController do
       @first  = FactoryGirl.create(:account, user: current_user, name: "The first one")
       @second = FactoryGirl.create(:account, user: current_user, name: "The second one")
 
-      get :index, query: "second"
+      get :index, params: { query: "second" }
       expect(assigns[:accounts]).to eq([@second])
       expect(assigns[:current_query]).to eq("second")
       expect(session[:accounts_current_query]).to eq("second")
@@ -64,7 +64,7 @@ describe AccountsController do
     describe "AJAX pagination" do
       it "should pick up page number from params" do
         @accounts = [FactoryGirl.create(:account, user: current_user)]
-        xhr :get, :index, page: 42
+        get :index, params: { page: 42 }, xhr: true
 
         expect(assigns[:current_page].to_i).to eq(42)
         expect(assigns[:accounts]).to eq([]) # page #42 should be empty if there's only one account ;-)
@@ -75,7 +75,7 @@ describe AccountsController do
       it "should pick up saved page number from session" do
         session[:accounts_current_page] = 42
         @accounts = [FactoryGirl.create(:account, user: current_user)]
-        xhr :get, :index
+        get :index, xhr: true
 
         expect(assigns[:current_page]).to eq(42)
         expect(assigns[:accounts]).to eq([])
@@ -86,7 +86,7 @@ describe AccountsController do
         session[:accounts_current_page] = 42
         session[:accounts_current_query] = "bill"
         @accounts = [FactoryGirl.create(:account, user: current_user)]
-        xhr :get, :index
+        get :index, xhr: true
 
         expect(assigns[:current_page]).to eq(1)
         expect(assigns[:accounts]).to eq(@accounts)
@@ -129,7 +129,7 @@ describe AccountsController do
       end
 
       it "should expose the requested account as @account and render [show] template" do
-        get :show, id: @account.id
+        get :show, params: { id: @account.id }
         expect(assigns[:account]).to eq(@account)
         expect(assigns[:stage]).to eq(@stage)
         expect(assigns[:comment].attributes).to eq(@comment.attributes)
@@ -137,7 +137,7 @@ describe AccountsController do
       end
 
       it "should update an activity when viewing the account" do
-        get :show, id: @account.id
+        get :show, params: { id: @account.id }
         expect(@account.versions.last.event).to eq('view')
       end
     end
@@ -149,7 +149,7 @@ describe AccountsController do
         expect(@account).to receive(:to_json).and_return("generated JSON")
 
         request.env["HTTP_ACCEPT"] = "application/json"
-        get :show, id: 42
+        get :show, params: { id: 42 }
         expect(response.body).to eq("generated JSON")
       end
     end
@@ -161,7 +161,7 @@ describe AccountsController do
         expect(@account).to receive(:to_xml).and_return("generated XML")
 
         request.env["HTTP_ACCEPT"] = "application/xml"
-        get :show, id: 42
+        get :show, params: { id: 42 }
         expect(response.body).to eq("generated XML")
       end
     end
@@ -171,7 +171,7 @@ describe AccountsController do
         @account = FactoryGirl.create(:account, user: current_user)
         @account.destroy
 
-        get :show, id: @account.id
+        get :show, params: { id: @account.id }
         expect(flash[:warning]).not_to eq(nil)
         expect(response).to redirect_to(accounts_path)
       end
@@ -179,7 +179,7 @@ describe AccountsController do
       it "should redirect to account index if the account is protected" do
         @private = FactoryGirl.create(:account, user: FactoryGirl.create(:user), access: "Private")
 
-        get :show, id: @private.id
+        get :show, params: { id: @private.id }
         expect(flash[:warning]).not_to eq(nil)
         expect(response).to redirect_to(accounts_path)
       end
@@ -189,7 +189,7 @@ describe AccountsController do
         @account.destroy
         request.env["HTTP_ACCEPT"] = "application/json"
 
-        get :show, id: @account.id
+        get :show, params: { id: @account.id }
         expect(response.code).to eq("404") # :not_found
       end
 
@@ -198,7 +198,7 @@ describe AccountsController do
         @account.destroy
         request.env["HTTP_ACCEPT"] = "application/xml"
 
-        get :show, id: @account.id
+        get :show, params: { id: @account.id }
         expect(response.code).to eq("404") # :not_found
       end
     end
@@ -211,7 +211,7 @@ describe AccountsController do
     it "should expose a new account as @account and render [new] template" do
       @account = Account.new(user: current_user,
                              access: Setting.default_access)
-      xhr :get, :new
+      get :new, xhr: true
       expect(assigns[:account].attributes).to eq(@account.attributes)
       expect(assigns[:contact]).to eq(nil)
       expect(response).to render_template("accounts/new")
@@ -220,7 +220,7 @@ describe AccountsController do
     it "should created an instance of related object when necessary" do
       @contact = FactoryGirl.create(:contact, id: 42)
 
-      xhr :get, :new, related: "contact_42"
+      get :new, params: { related: "contact_42" }, xhr: true
       expect(assigns[:contact]).to eq(@contact)
     end
   end
@@ -231,7 +231,7 @@ describe AccountsController do
     it "should expose the requested account as @account and render [edit] template" do
       @account = FactoryGirl.create(:account, id: 42, user: current_user)
 
-      xhr :get, :edit, id: 42
+      get :edit, params: { id: 42 }, xhr: true
       expect(assigns[:account]).to eq(@account)
       expect(assigns[:previous]).to eq(nil)
       expect(response).to render_template("accounts/edit")
@@ -241,7 +241,7 @@ describe AccountsController do
       @account = FactoryGirl.create(:account, id: 42)
       @previous = FactoryGirl.create(:account, id: 41)
 
-      xhr :get, :edit, id: 42, previous: 41
+      get :edit, params: { id: 42, previous: 41 }, xhr: true
       expect(assigns[:previous]).to eq(@previous)
     end
 
@@ -250,7 +250,7 @@ describe AccountsController do
         @account = FactoryGirl.create(:account, user: current_user)
         @account.destroy
 
-        xhr :get, :edit, id: @account.id
+        get :edit, params: { id: @account.id }, xhr: true
         expect(flash[:warning]).not_to eq(nil)
         expect(response.body).to eq("window.location.reload();")
       end
@@ -258,7 +258,7 @@ describe AccountsController do
       it "should reload current page with the flash message if the account is protected" do
         @private = FactoryGirl.create(:account, user: FactoryGirl.create(:user), access: "Private")
 
-        xhr :get, :edit, id: @private.id
+        get :edit, params: { id: @private.id }, xhr: true
         expect(flash[:warning]).not_to eq(nil)
         expect(response.body).to eq("window.location.reload();")
       end
@@ -273,7 +273,7 @@ describe AccountsController do
       it "should notify the view if previous account got deleted" do
         @previous.destroy
 
-        xhr :get, :edit, id: @account.id, previous: @previous.id
+        get :edit, params: { id: @account.id, previous: @previous.id }, xhr: true
         expect(flash[:warning]).to eq(nil) # no warning, just silently remove the div
         expect(assigns[:previous]).to eq(@previous.id)
         expect(response).to render_template("accounts/edit")
@@ -282,7 +282,7 @@ describe AccountsController do
       it "should notify the view if previous account got protected" do
         @previous.update_attribute(:access, "Private")
 
-        xhr :get, :edit, id: @account.id, previous: @previous.id
+        get :edit, params: { id: @account.id, previous: @previous.id }, xhr: true
         expect(flash[:warning]).to eq(nil)
         expect(assigns[:previous]).to eq(@previous.id)
         expect(response).to render_template("accounts/edit")
@@ -299,7 +299,7 @@ describe AccountsController do
         @account = FactoryGirl.build(:account, name: "Hello world", user: current_user)
         allow(Account).to receive(:new).and_return(@account)
 
-        xhr :post, :create, account: { name: "Hello world" }
+        post :create, params: { account: { name: "Hello world" } }, xhr: true
         expect(assigns(:account)).to eq(@account)
         expect(response).to render_template("accounts/create")
       end
@@ -309,7 +309,7 @@ describe AccountsController do
         @account = FactoryGirl.build(:account, user: current_user)
         allow(Account).to receive(:new).and_return(@account)
 
-        xhr :post, :create, account: { name: "Hello" }
+        post :create, params: { account: { name: "Hello" } }, xhr: true
         expect(assigns[:accounts]).to eq([@account])
       end
 
@@ -317,7 +317,7 @@ describe AccountsController do
         @account = FactoryGirl.build(:account, name: "Hello", user: current_user)
         allow(Campaign).to receive(:new).and_return(@account)
 
-        xhr :post, :create, account: { name: "Hello" }
+        post :create, params: { account: { name: "Hello" } }, xhr: true
         expect(assigns[:account_category_total]).to be_instance_of(HashWithIndifferentAccess)
       end
 
@@ -325,7 +325,7 @@ describe AccountsController do
         @account = FactoryGirl.build(:account, name: "Hello world", user: current_user)
         allow(Account).to receive(:new).and_return(@account)
 
-        xhr :post, :create, account: { name: "Hello world" }, comment_body: "Awesome comment is awesome"
+        post :create, params: { account: { name: "Hello world" }, comment_body: "Awesome comment is awesome" }, xhr: true
         expect(assigns[:account].comments.map(&:comment)).to include("Awesome comment is awesome")
       end
     end
@@ -335,7 +335,7 @@ describe AccountsController do
         @account = FactoryGirl.build(:account, name: nil, user: nil)
         allow(Account).to receive(:new).and_return(@account)
 
-        xhr :post, :create, account: {}
+        post :create, params: { account: {} }, xhr: true
         expect(assigns(:account)).to eq(@account)
         expect(response).to render_template("accounts/create")
       end
@@ -350,7 +350,7 @@ describe AccountsController do
       it "should update the requested account, expose the requested account as @account, and render [update] template" do
         @account = FactoryGirl.create(:account, id: 42, name: "Hello people")
 
-        xhr :put, :update, id: 42, account: { name: "Hello world" }
+        put :update, params: { id: 42, account: { name: "Hello world" } }, xhr: true
         expect(@account.reload.name).to eq("Hello world")
         expect(assigns(:account)).to eq(@account)
         expect(response).to render_template("accounts/update")
@@ -360,7 +360,7 @@ describe AccountsController do
         @account = FactoryGirl.create(:account, id: 42)
         request.env["HTTP_REFERER"] = "http://localhost/accounts"
 
-        xhr :put, :update, id: 42, account: { name: "Hello" }
+        put :update, params: { id: 42, account: { name: "Hello" } }, xhr: true
         expect(assigns(:account)).to eq(@account)
         expect(assigns[:account_category_total]).to be_instance_of(HashWithIndifferentAccess)
       end
@@ -368,7 +368,7 @@ describe AccountsController do
       it "should update account permissions when sharing with specific users" do
         @account = FactoryGirl.create(:account, id: 42, access: "Public")
 
-        xhr :put, :update, id: 42, account: { name: "Hello", access: "Shared", user_ids: [7, 8] }
+        put :update, params: { id: 42, account: { name: "Hello", access: "Shared", user_ids: [7, 8] } }, xhr: true
         expect(assigns[:account].access).to eq("Shared")
         expect(assigns[:account].user_ids.sort).to eq([7, 8])
       end
@@ -378,7 +378,7 @@ describe AccountsController do
           @account = FactoryGirl.create(:account, user: current_user)
           @account.destroy
 
-          xhr :put, :update, id: @account.id
+          put :update, params: { id: @account.id }, xhr: true
           expect(flash[:warning]).not_to eq(nil)
           expect(response.body).to eq("window.location.reload();")
         end
@@ -386,7 +386,7 @@ describe AccountsController do
         it "should reload current page with the flash message if the account is protected" do
           @private = FactoryGirl.create(:account, user: FactoryGirl.create(:user), access: "Private")
 
-          xhr :put, :update, id: @private.id
+          put :update, params: { id: @private.id }, xhr: true
           expect(flash[:warning]).not_to eq(nil)
           expect(response.body).to eq("window.location.reload();")
         end
@@ -397,7 +397,7 @@ describe AccountsController do
       it "should not update the requested account but still expose the requested account as @account, and render [update] template" do
         @account = FactoryGirl.create(:account, id: 42, name: "Hello people")
 
-        xhr :put, :update, id: 42, account: { name: nil }
+        put :update, params: { id: 42, account: { name: nil } }, xhr: true
         expect(assigns(:account).reload.name).to eq("Hello people")
         expect(assigns(:account)).to eq(@account)
         expect(response).to render_template("accounts/update")
@@ -416,7 +416,7 @@ describe AccountsController do
     describe "AJAX request" do
       it "should destroy the requested account and render [destroy] template" do
         @another_account = FactoryGirl.create(:account, user: current_user)
-        xhr :delete, :destroy, id: @account.id
+        delete :destroy, params: { id: @account.id }, xhr: true
 
         expect { Account.find(@account.id) }.to raise_error(ActiveRecord::RecordNotFound)
         expect(assigns[:accounts]).to eq([@another_account]) # @account got deleted
@@ -424,7 +424,7 @@ describe AccountsController do
       end
 
       it "should get data for accounts sidebar" do
-        xhr :delete, :destroy, id: @account.id
+        delete :destroy, params: { id: @account.id }, xhr: true
 
         expect(assigns[:account_category_total]).to be_instance_of(HashWithIndifferentAccess)
       end
@@ -432,7 +432,7 @@ describe AccountsController do
       it "should try previous page and render index action if current page has no accounts" do
         session[:accounts_current_page] = 42
 
-        xhr :delete, :destroy, id: @account.id
+        delete :destroy, params: { id: @account.id }, xhr: true
         expect(session[:accounts_current_page]).to eq(41)
         expect(response).to render_template("accounts/index")
       end
@@ -440,7 +440,7 @@ describe AccountsController do
       it "should render index action when deleting last account" do
         session[:accounts_current_page] = 1
 
-        xhr :delete, :destroy, id: @account.id
+        delete :destroy, params: { id: @account.id }, xhr: true
         expect(session[:accounts_current_page]).to eq(1)
         expect(response).to render_template("accounts/index")
       end
@@ -450,7 +450,7 @@ describe AccountsController do
           @account = FactoryGirl.create(:account, user: current_user)
           @account.destroy
 
-          xhr :delete, :destroy, id: @account.id
+          delete :destroy, params: { id: @account.id }, xhr: true
           expect(flash[:warning]).not_to eq(nil)
           expect(response.body).to eq("window.location.reload();")
         end
@@ -458,7 +458,7 @@ describe AccountsController do
         it "should reload current page with the flash message if the account is protected" do
           @private = FactoryGirl.create(:account, user: FactoryGirl.create(:user), access: "Private")
 
-          xhr :delete, :destroy, id: @private.id
+          delete :destroy, params: { id: @private.id }, xhr: true
           expect(flash[:warning]).not_to eq(nil)
           expect(response.body).to eq("window.location.reload();")
         end
@@ -467,7 +467,7 @@ describe AccountsController do
 
     describe "HTML request" do
       it "should redirect to Accounts index when an account gets deleted from its landing page" do
-        delete :destroy, id: @account.id
+        delete :destroy, params: { id: @account.id }
 
         expect(flash[:notice]).not_to eq(nil)
         expect(response).to redirect_to(accounts_path)
@@ -477,7 +477,7 @@ describe AccountsController do
         @account = FactoryGirl.create(:account, user: current_user)
         @account.destroy
 
-        delete :destroy, id: @account.id
+        delete :destroy, params: { id: @account.id }
         expect(flash[:warning]).not_to eq(nil)
         expect(response).to redirect_to(accounts_path)
       end
@@ -485,7 +485,7 @@ describe AccountsController do
       it "should redirect to account index with the flash message if the account is protected" do
         @private = FactoryGirl.create(:account, user: FactoryGirl.create(:user), access: "Private")
 
-        delete :destroy, id: @private.id
+        delete :destroy, params: { id: @private.id }
         expect(flash[:warning]).not_to eq(nil)
         expect(response).to redirect_to(accounts_path)
       end
@@ -560,14 +560,14 @@ describe AccountsController do
   #----------------------------------------------------------------------------
   describe "responding to GET redraw" do
     it "should save user selected account preference" do
-      xhr :get, :redraw, per_page: 42, view: "brief", sort_by: "name"
+      get :redraw, params: { per_page: 42, view: "brief", sort_by: "name" }, xhr: true
       expect(current_user.preference[:accounts_per_page]).to eq("42")
       expect(current_user.preference[:accounts_index_view]).to eq("brief")
       expect(current_user.preference[:accounts_sort_by]).to eq("accounts.name ASC")
     end
 
     it "should reset current page to 1" do
-      xhr :get, :redraw, per_page: 42, view: "brief", sort_by: "name"
+      get :redraw, params: { per_page: 42, view: "brief", sort_by: "name" }, xhr: true
       expect(session[:accounts_current_page]).to eq(1)
     end
 
@@ -577,7 +577,7 @@ describe AccountsController do
         FactoryGirl.create(:account, name: "B", user: current_user)
       ]
 
-      xhr :get, :redraw, per_page: 1, sort_by: "name"
+      get :redraw, params: { per_page: 1, sort_by: "name" }, xhr: true
       expect(assigns(:accounts)).to eq([@accounts.first])
       expect(response).to render_template("accounts/index")
     end
@@ -590,14 +590,14 @@ describe AccountsController do
       session[:accounts_filter] = "customer,vendor"
       @accounts = [FactoryGirl.create(:account, category: "partner", user: current_user)]
 
-      xhr :post, :filter, category: "partner"
+      post :filter, params: { category: "partner" }, xhr: true
       expect(assigns(:accounts)).to eq(@accounts)
       expect(response).to render_template("accounts/index")
     end
 
     it "should reset current page to 1" do
       @accounts = []
-      xhr :post, :filter, category: "partner"
+      post :filter, params: { category: "partner" }, xhr: true
 
       expect(session[:accounts_current_page]).to eq(1)
     end
