@@ -51,7 +51,7 @@ describe LeadsController do
       @billy_bones   = FactoryGirl.create(:lead, user: current_user, first_name: "Billy",   last_name: "Bones")
       @captain_flint = FactoryGirl.create(:lead, user: current_user, first_name: "Captain", last_name: "Flint")
 
-      get :index, query: "bill"
+      get :index, params: { query: "bill" }
       expect(assigns[:leads]).to eq([@billy_bones])
       expect(assigns[:current_query]).to eq("bill")
       expect(session[:leads_current_query]).to eq("bill")
@@ -60,7 +60,7 @@ describe LeadsController do
     describe "AJAX pagination" do
       it "should pick up page number from params" do
         @leads = [FactoryGirl.create(:lead, user: current_user)]
-        xhr :get, :index, page: 42
+        get :index, params: { page: 42 }, xhr: true
 
         expect(assigns[:current_page].to_i).to eq(42)
         expect(assigns[:leads]).to eq([]) # page #42 should be empty if there's only one lead ;-)
@@ -72,7 +72,7 @@ describe LeadsController do
         session[:leads_current_page] = 42
         session[:leads_current_query] = "bill"
         @leads = [FactoryGirl.create(:lead, user: current_user)]
-        xhr :get, :index, query: "bill"
+        get :index, params: { query: "bill" }, xhr: true
 
         expect(assigns[:current_page]).to eq(42)
         expect(assigns[:leads]).to eq([])
@@ -83,7 +83,7 @@ describe LeadsController do
         session[:leads_current_page] = 42
         session[:leads_current_query] = "bill"
         @leads = [FactoryGirl.create(:lead, user: current_user)]
-        xhr :get, :index
+        get :index, xhr: true
 
         expect(assigns[:current_page]).to eq(1)
         expect(assigns[:leads]).to eq(@leads)
@@ -125,14 +125,14 @@ describe LeadsController do
       end
 
       it "should expose the requested lead as @lead and render [show] template" do
-        get :show, id: 42
+        get :show, params: { id: 42 }
         expect(assigns[:lead]).to eq(@lead)
         expect(assigns[:comment].attributes).to eq(@comment.attributes)
         expect(response).to render_template("leads/show")
       end
 
       it "should update an activity when viewing the lead" do
-        get :show, id: @lead.id
+        get :show, params: { id: @lead.id }
         expect(@lead.versions.last.event).to eq('view')
       end
     end
@@ -144,7 +144,7 @@ describe LeadsController do
         expect(@lead).to receive(:to_json).and_return("generated JSON")
 
         request.env["HTTP_ACCEPT"] = "application/json"
-        get :show, id: 42
+        get :show, params: { id: 42 }
         expect(response.body).to eq("generated JSON")
       end
     end
@@ -156,7 +156,7 @@ describe LeadsController do
         expect(@lead).to receive(:to_xml).and_return("generated XML")
 
         request.env["HTTP_ACCEPT"] = "application/xml"
-        get :show, id: 42
+        get :show, params: { id: 42 }
         expect(response.body).to eq("generated XML")
       end
     end
@@ -166,7 +166,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, user: current_user)
         @lead.destroy
 
-        get :show, id: @lead.id
+        get :show, params: { id: @lead.id }
         expect(flash[:warning]).not_to eq(nil)
         expect(response).to redirect_to(leads_path)
       end
@@ -174,7 +174,7 @@ describe LeadsController do
       it "should redirect to lead index if the lead is protected" do
         @private = FactoryGirl.create(:lead, user: FactoryGirl.create(:user), access: "Private")
 
-        get :show, id: @private.id
+        get :show, params: { id: @private.id }
         expect(flash[:warning]).not_to eq(nil)
         expect(response).to redirect_to(leads_path)
       end
@@ -184,7 +184,7 @@ describe LeadsController do
         @lead.destroy
         request.env["HTTP_ACCEPT"] = "application/json"
 
-        get :show, id: @lead.id
+        get :show, params: { id: @lead.id }
         expect(response.code).to eq("404") # :not_found
       end
 
@@ -193,7 +193,7 @@ describe LeadsController do
         @lead.destroy
         request.env["HTTP_ACCEPT"] = "application/xml"
 
-        get :show, id: @lead.id
+        get :show, params: { id: @lead.id }
         expect(response.code).to eq("404") # :not_found
       end
     end
@@ -208,7 +208,7 @@ describe LeadsController do
       allow(Lead).to receive(:new).and_return(@lead)
       @campaigns = [FactoryGirl.create(:campaign, user: current_user)]
 
-      xhr :get, :new
+      get :new, xhr: true
       expect(assigns[:lead].attributes).to eq(@lead.attributes)
       expect(assigns[:campaigns]).to eq(@campaigns)
       expect(response).to render_template("leads/new")
@@ -217,7 +217,7 @@ describe LeadsController do
     it "should create related object when necessary" do
       @campaign = FactoryGirl.create(:campaign, id: 123)
 
-      xhr :get, :new, related: "campaign_123"
+      get :new, params: { related: "campaign_123" }, xhr: true
       expect(assigns[:campaign]).to eq(@campaign)
     end
 
@@ -226,7 +226,7 @@ describe LeadsController do
         @campaign = FactoryGirl.create(:campaign)
         @campaign.destroy
 
-        xhr :get, :new, related: "campaign_#{@campaign.id}"
+        get :new, params: { related: "campaign_#{@campaign.id}" }, xhr: true
         expect(flash[:warning]).not_to eq(nil)
         expect(response.body).to eq('window.location.href = "/campaigns";')
       end
@@ -234,7 +234,7 @@ describe LeadsController do
       it "should redirect to parent asset's index page with the message if parent asset got protected" do
         @campaign = FactoryGirl.create(:campaign, access: "Private")
 
-        xhr :get, :new, related: "campaign_#{@campaign.id}"
+        get :new, params: { related: "campaign_#{@campaign.id}" }, xhr: true
         expect(flash[:warning]).not_to eq(nil)
         expect(response.body).to eq('window.location.href = "/campaigns";')
       end
@@ -248,7 +248,7 @@ describe LeadsController do
       @lead = FactoryGirl.create(:lead, id: 42, user: current_user, campaign: nil)
       @campaigns = [FactoryGirl.create(:campaign, user: current_user)]
 
-      xhr :get, :edit, id: 42
+      get :edit, params: { id: 42 }, xhr: true
       expect(assigns[:lead]).to eq(@lead)
       expect(assigns[:campaigns]).to eq(@campaigns)
       expect(response).to render_template("leads/edit")
@@ -258,7 +258,7 @@ describe LeadsController do
       @lead = FactoryGirl.create(:lead, id: 42)
       @previous = FactoryGirl.create(:lead, id: 321)
 
-      xhr :get, :edit, id: 42, previous: 321
+      get :edit, params: { id: 42, previous: 321 }, xhr: true
       expect(assigns[:previous]).to eq(@previous)
     end
 
@@ -267,7 +267,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, user: current_user)
         @lead.destroy
 
-        xhr :get, :edit, id: @lead.id
+        get :edit, params: { id: @lead.id }, xhr: true
         expect(flash[:warning]).not_to eq(nil)
         expect(response.body).to eq("window.location.reload();")
       end
@@ -275,7 +275,7 @@ describe LeadsController do
       it "should reload current page with the flash message if the lead is protected" do
         @private = FactoryGirl.create(:lead, user: FactoryGirl.create(:user), access: "Private")
 
-        xhr :get, :edit, id: @private.id
+        get :edit, params: { id: @private.id }, xhr: true
         expect(flash[:warning]).not_to eq(nil)
         expect(response.body).to eq("window.location.reload();")
       end
@@ -290,7 +290,7 @@ describe LeadsController do
       it "should notify the view if previous lead got deleted" do
         @previous.destroy
 
-        xhr :get, :edit, id: @lead.id, previous: @previous.id
+        get :edit, params: { id: @lead.id, previous: @previous.id }, xhr: true
         expect(flash[:warning]).to eq(nil) # no warning, just silently remove the div
         expect(assigns[:previous]).to eq(@previous.id)
         expect(response).to render_template("leads/edit")
@@ -299,7 +299,7 @@ describe LeadsController do
       it "should notify the view if previous lead got protected" do
         @previous.update_attribute(:access, "Private")
 
-        xhr :get, :edit, id: @lead.id, previous: @previous.id
+        get :edit, params: { id: @lead.id, previous: @previous.id }, xhr: true
         expect(flash[:warning]).to eq(nil)
         expect(assigns[:previous]).to eq(@previous.id)
         expect(response).to render_template("leads/edit")
@@ -317,7 +317,7 @@ describe LeadsController do
         allow(Lead).to receive(:new).and_return(@lead)
         @campaigns = [FactoryGirl.create(:campaign, user: current_user)]
 
-        xhr :post, :create, lead: { first_name: "Billy", last_name: "Bones" }
+        post :create, params: { lead: { first_name: "Billy", last_name: "Bones" } }, xhr: true
         expect(assigns(:lead)).to eq(@lead)
         expect(assigns(:campaigns)).to eq(@campaigns)
         expect(assigns[:lead_status_total]).to be_nil
@@ -335,7 +335,7 @@ describe LeadsController do
         @lead = FactoryGirl.build(:lead, campaign: @campaign, user: current_user, access: "Shared")
         allow(Lead).to receive(:new).and_return(@lead)
 
-        xhr :post, :create, lead: { first_name: "Billy", last_name: "Bones", access: "Campaign", user_ids: %w(7 8) }, campaign: @campaign.id
+        post :create, params: { lead: { first_name: "Billy", last_name: "Bones", access: "Campaign", user_ids: %w(7 8) }, campaign: @campaign.id }, xhr: true
         expect(assigns(:lead)).to eq(@lead)
         expect(@lead.reload.access).to eq("Shared")
         expect(@lead.permissions.map(&:user_id).sort).to eq([7, 8])
@@ -348,7 +348,7 @@ describe LeadsController do
         allow(Lead).to receive(:new).and_return(@lead)
 
         request.env["HTTP_REFERER"] = "http://localhost/leads"
-        xhr :post, :create, lead: { first_name: "Billy", last_name: "Bones" }
+        post :create, params: { lead: { first_name: "Billy", last_name: "Bones" } }, xhr: true
         expect(assigns[:lead_status_total]).to be_an_instance_of(HashWithIndifferentAccess)
       end
 
@@ -357,7 +357,7 @@ describe LeadsController do
         allow(Lead).to receive(:new).and_return(@lead)
 
         request.env["HTTP_REFERER"] = "http://localhost/leads"
-        xhr :post, :create, lead: { first_name: "Billy", last_name: "Bones" }
+        post :create, params: { lead: { first_name: "Billy", last_name: "Bones" } }, xhr: true
         expect(assigns[:leads]).to eq([@lead])
       end
 
@@ -366,14 +366,14 @@ describe LeadsController do
         @lead = FactoryGirl.build(:lead, user: current_user, campaign: @campaign)
 
         request.env["HTTP_REFERER"] = "http://localhost/campaigns/#{@campaign.id}"
-        xhr :put, :create, lead: { first_name: "Billy", last_name: "Bones" }, campaign: @campaign.id
+        put :create, params: { lead: { first_name: "Billy", last_name: "Bones" }, campaign: @campaign.id }, xhr: true
         expect(assigns[:campaign]).to eq(@campaign)
       end
 
       it "should add a new comment to the newly created lead when specified" do
         @lead = FactoryGirl.create(:lead)
         allow(Lead).to receive(:new).and_return(@lead)
-        xhr :post, :create, lead: { first_name: "Test", last_name: "Lead" }, comment_body: "This is an important lead."
+        post :create, params: { lead: { first_name: "Test", last_name: "Lead" }, comment_body: "This is an important lead." }, xhr: true
         expect(@lead.reload.comments.map(&:comment)).to include("This is an important lead.")
       end
     end
@@ -384,7 +384,7 @@ describe LeadsController do
         allow(Lead).to receive(:new).and_return(@lead)
         @campaigns = [FactoryGirl.create(:campaign, user: current_user)]
 
-        xhr :post, :create, lead: { first_name: nil }
+        post :create, params: { lead: { first_name: nil } }, xhr: true
         expect(assigns(:lead)).to eq(@lead)
         expect(assigns(:campaigns)).to eq(@campaigns)
         expect(assigns[:lead_status_total]).to eq(nil)
@@ -401,7 +401,7 @@ describe LeadsController do
       it "should update the requested lead, expose it as @lead, and render [update] template" do
         @lead = FactoryGirl.create(:lead, first_name: "Billy", user: current_user)
 
-        xhr :put, :update, id: @lead.id, lead: { first_name: "Bones" }
+        put :update, params: { id: @lead.id, lead: { first_name: "Bones" } }, xhr: true
         expect(@lead.reload.first_name).to eq("Bones")
         expect(assigns[:lead]).to eq(@lead)
         expect(assigns[:lead_status_total]).to eq(nil)
@@ -411,14 +411,14 @@ describe LeadsController do
       it "should update lead status" do
         @lead = FactoryGirl.create(:lead, status: "new", user: current_user)
 
-        xhr :put, :update, id: @lead.id, lead: { status: "rejected" }
+        put :update, params: { id: @lead.id, lead: { status: "rejected" } }, xhr: true
         expect(@lead.reload.status).to eq("rejected")
       end
 
       it "should update lead source" do
         @lead = FactoryGirl.create(:lead, source: "campaign", user: current_user)
 
-        xhr :put, :update, id: @lead.id, lead: { source: "cald_call" }
+        put :update, params: { id: @lead.id, lead: { source: "cald_call" } }, xhr: true
         expect(@lead.reload.source).to eq("cald_call")
       end
 
@@ -426,7 +426,7 @@ describe LeadsController do
         @campaigns = { old: FactoryGirl.create(:campaign), new: FactoryGirl.create(:campaign) }
         @lead = FactoryGirl.create(:lead, campaign: @campaigns[:old])
 
-        xhr :put, :update, id: @lead.id, lead: { campaign_id: @campaigns[:new].id }
+        put :update, params: { id: @lead.id, lead: { campaign_id: @campaigns[:new].id } }, xhr: true
         expect(@lead.reload.campaign).to eq(@campaigns[:new])
       end
 
@@ -435,7 +435,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, campaign: @campaign)
         @count = @campaign.reload.leads_count
 
-        xhr :put, :update, id: @lead, lead: { campaign_id: nil }
+        put :update, params: { id: @lead, lead: { campaign_id: nil } }, xhr: true
         expect(@lead.reload.campaign).to eq(nil)
         expect(@campaign.reload.leads_count).to eq(@count - 1)
       end
@@ -445,7 +445,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, campaign: nil)
         @count = @campaign.leads_count
 
-        xhr :put, :update, id: @lead, lead: { campaign_id: @campaign.id }
+        put :update, params: { id: @lead, lead: { campaign_id: @campaign.id } }, xhr: true
         expect(@lead.reload.campaign).to eq(@campaign)
         expect(@campaign.reload.leads_count).to eq(@count + 1)
       end
@@ -455,7 +455,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, campaign: @campaigns[:old])
         @counts = { old: @campaigns[:old].reload.leads_count, new: @campaigns[:new].leads_count }
 
-        xhr :put, :update, id: @lead, lead: { campaign_id: @campaigns[:new].id }
+        put :update, params: { id: @lead, lead: { campaign_id: @campaigns[:new].id } }, xhr: true
         expect(@lead.reload.campaign).to eq(@campaigns[:new])
         expect(@campaigns[:old].reload.leads_count).to eq(@counts[:old] - 1)
         expect(@campaigns[:new].reload.leads_count).to eq(@counts[:new] + 1)
@@ -466,7 +466,7 @@ describe LeadsController do
         he  = FactoryGirl.create(:user, id: 7)
         she = FactoryGirl.create(:user, id: 8)
 
-        xhr :put, :update, id: @lead.id, lead: { access: "Shared", user_ids: %w(7 8) }
+        put :update, params: { id: @lead.id, lead: { access: "Shared", user_ids: %w(7 8) } }, xhr: true
         expect(@lead.user_ids.sort).to eq([he.id, she.id])
       end
 
@@ -474,7 +474,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead)
 
         request.env["HTTP_REFERER"] = "http://localhost/leads"
-        xhr :put, :update, id: @lead.id, lead: { first_name: "Billy" }
+        put :update, params: { id: @lead.id, lead: { first_name: "Billy" } }, xhr: true
         expect(assigns[:lead_status_total]).not_to be_nil
         expect(assigns[:lead_status_total]).to be_an_instance_of(HashWithIndifferentAccess)
       end
@@ -484,7 +484,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, campaign: @campaign)
 
         request.env["HTTP_REFERER"] = "http://localhost/campaigns/#{@campaign.id}"
-        xhr :put, :update, id: @lead.id, lead: { first_name: "Hello" }
+        put :update, params: { id: @lead.id, lead: { first_name: "Hello" } }, xhr: true
         expect(assigns[:campaign]).to eq(@campaign)
       end
 
@@ -493,7 +493,7 @@ describe LeadsController do
           @lead = FactoryGirl.create(:lead, user: current_user)
           @lead.destroy
 
-          xhr :put, :update, id: @lead.id
+          put :update, params: { id: @lead.id }, xhr: true
           expect(flash[:warning]).not_to eq(nil)
           expect(response.body).to eq("window.location.reload();")
         end
@@ -501,7 +501,7 @@ describe LeadsController do
         it "should reload current page with the flash message if the lead is protected" do
           @private = FactoryGirl.create(:lead, user: FactoryGirl.create(:user), access: "Private")
 
-          xhr :put, :update, id: @private.id
+          put :update, params: { id: @private.id }, xhr: true
           expect(flash[:warning]).not_to eq(nil)
           expect(response.body).to eq("window.location.reload();")
         end
@@ -513,7 +513,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, id: 42, user: current_user, campaign: nil)
         @campaigns = [FactoryGirl.create(:campaign, user: current_user)]
 
-        xhr :put, :update, id: 42, lead: { first_name: nil }
+        put :update, params: { id: 42, lead: { first_name: nil } }, xhr: true
         expect(assigns[:lead]).to eq(@lead)
         expect(assigns[:campaigns]).to eq(@campaigns)
         expect(response).to render_template("leads/update")
@@ -531,7 +531,7 @@ describe LeadsController do
 
     describe "AJAX request" do
       it "should destroy the requested lead and render [destroy] template" do
-        xhr :delete, :destroy, id: @lead.id
+        delete :destroy, params: { id: @lead.id }, xhr: true
 
         expect(assigns[:leads]).to eq(nil) # @lead got deleted
         expect { Lead.find(@lead.id) }.to raise_error(ActiveRecord::RecordNotFound)
@@ -546,7 +546,7 @@ describe LeadsController do
         it "should get data for the sidebar" do
           @another_lead = FactoryGirl.create(:lead, user: current_user)
 
-          xhr :delete, :destroy, id: @lead.id
+          delete :destroy, params: { id: @lead.id }, xhr: true
           expect(assigns[:leads]).to eq([@another_lead]) # @lead got deleted
           expect(assigns[:lead_status_total]).not_to be_nil
           expect(assigns[:lead_status_total]).to be_an_instance_of(HashWithIndifferentAccess)
@@ -556,7 +556,7 @@ describe LeadsController do
         it "should try previous page and render index action if current page has no leads" do
           session[:leads_current_page] = 42
 
-          xhr :delete, :destroy, id: @lead.id
+          delete :destroy, params: { id: @lead.id }, xhr: true
           expect(session[:leads_current_page]).to eq(41)
           expect(response).to render_template("leads/index")
         end
@@ -564,7 +564,7 @@ describe LeadsController do
         it "should render index action when deleting last lead" do
           session[:leads_current_page] = 1
 
-          xhr :delete, :destroy, id: @lead.id
+          delete :destroy, params: { id: @lead.id }, xhr: true
           expect(session[:leads_current_page]).to eq(1)
           expect(response).to render_template("leads/index")
         end
@@ -578,13 +578,13 @@ describe LeadsController do
         end
 
         it "should reset current page to 1" do
-          xhr :delete, :destroy, id: @lead.id
+          delete :destroy, params: { id: @lead.id }, xhr: true
           expect(session[:leads_current_page]).to eq(1)
           expect(response).to render_template("leads/destroy")
         end
 
         it "should reload campaiign to be able to refresh its summary" do
-          xhr :delete, :destroy, id: @lead.id
+          delete :destroy, params: { id: @lead.id }, xhr: true
           expect(assigns[:campaign]).to eq(@campaign)
           expect(response).to render_template("leads/destroy")
         end
@@ -595,7 +595,7 @@ describe LeadsController do
           @lead = FactoryGirl.create(:lead, user: current_user)
           @lead.destroy
 
-          xhr :delete, :destroy, id: @lead.id
+          delete :destroy, params: { id: @lead.id }, xhr: true
           expect(flash[:warning]).not_to eq(nil)
           expect(response.body).to eq("window.location.reload();")
         end
@@ -603,7 +603,7 @@ describe LeadsController do
         it "should reload current page with the flash message if the lead is protected" do
           @private = FactoryGirl.create(:lead, user: FactoryGirl.create(:user), access: "Private")
 
-          xhr :delete, :destroy, id: @private.id
+          delete :destroy, params: { id: @private.id }, xhr: true
           expect(flash[:warning]).not_to eq(nil)
           expect(response.body).to eq("window.location.reload();")
         end
@@ -612,7 +612,7 @@ describe LeadsController do
 
     describe "HTML request" do
       it "should redirect to Leads index when a lead gets deleted from its landing page" do
-        delete :destroy, id: @lead.id
+        delete :destroy, params: { id: @lead.id }
         expect(flash[:notice]).not_to eq(nil)
         expect(session[:leads_current_page]).to eq(1)
         expect(response).to redirect_to(leads_path)
@@ -622,7 +622,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, user: current_user)
         @lead.destroy
 
-        delete :destroy, id: @lead.id
+        delete :destroy, params: { id: @lead.id }
         expect(flash[:warning]).not_to eq(nil)
         expect(response).to redirect_to(leads_path)
       end
@@ -630,7 +630,7 @@ describe LeadsController do
       it "should redirect to lead index with the flash message if the lead is protected" do
         @private = FactoryGirl.create(:lead, user: FactoryGirl.create(:user), access: "Private")
 
-        delete :destroy, id: @private.id
+        delete :destroy, params: { id: @private.id }
         expect(flash[:warning]).not_to eq(nil)
         expect(response).to redirect_to(leads_path)
       end
@@ -648,7 +648,7 @@ describe LeadsController do
       @account = Account.new(user: current_user, name: @lead.company, access: "Lead")
       @opportunity = Opportunity.new(user: current_user, access: "Lead", stage: "prospecting", campaign: @lead.campaign, source: @lead.source)
 
-      xhr :get, :convert, id: @lead.id
+      get :convert, params: { id: @lead.id }, xhr: true
       expect(assigns[:lead]).to eq(@lead)
       expect(assigns[:accounts]).to eq(@accounts)
       expect(assigns[:account].attributes).to eq(@account.attributes)
@@ -662,7 +662,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, user: current_user)
         @lead.destroy
 
-        xhr :get, :convert, id: @lead.id
+        get :convert, params: { id: @lead.id }, xhr: true
         expect(flash[:warning]).not_to eq(nil)
         expect(response.body).to eq("window.location.reload();")
       end
@@ -670,7 +670,7 @@ describe LeadsController do
       it "should reload current page with the flash message if the lead is protected" do
         @private = FactoryGirl.create(:lead, user: FactoryGirl.create(:user), access: "Private")
 
-        xhr :get, :convert, id: @private.id
+        get :convert, params: { id: @private.id }, xhr: true
         expect(flash[:warning]).not_to eq(nil)
         expect(response.body).to eq("window.location.reload();")
       end
@@ -685,7 +685,7 @@ describe LeadsController do
       it "should notify the view if previous lead got deleted" do
         @previous.destroy
 
-        xhr :get, :convert, id: @lead.id, previous: @previous.id
+        get :convert, params: { id: @lead.id, previous: @previous.id }, xhr: true
         expect(flash[:warning]).to eq(nil) # no warning, just silently remove the div
         expect(assigns[:previous]).to eq(@previous.id)
         expect(response).to render_template("leads/convert")
@@ -694,7 +694,7 @@ describe LeadsController do
       it "should notify the view if previous lead got protected" do
         @previous.update_attribute(:access, "Private")
 
-        xhr :get, :convert, id: @lead.id, previous: @previous.id
+        get :convert, params: { id: @lead.id, previous: @previous.id }, xhr: true
         expect(flash[:warning]).to eq(nil)
         expect(assigns[:previous]).to eq(@previous.id)
         expect(response).to render_template("leads/convert")
@@ -715,7 +715,7 @@ describe LeadsController do
       @contact = FactoryGirl.build(:contact, user: current_user, lead: @lead)
       allow(Contact).to receive(:new).and_return(@contact)
 
-      xhr :put, :promote, id: 42, account: { id: 123 }, opportunity: { name: "Hello" }
+      put :promote, params: { id: 42, account: { id: 123 }, opportunity: { name: "Hello" } }, xhr: true
       expect(@lead.reload.status).to eq("converted")
       expect(assigns[:lead]).to eq(@lead)
       expect(assigns[:account]).to eq(@account)
@@ -742,7 +742,7 @@ describe LeadsController do
       @opportunity.permissions << FactoryGirl.create(:permission, user: she, asset: @opportunity)
       allow(@opportunity).to receive(:new).and_return(@opportunity)
 
-      xhr :put, :promote, id: @lead.id, access: "Lead", account: { name: "Hello", access: "Lead", user_id: current_user.id }, opportunity: { name: "World", access: "Lead", user_id: current_user.id }
+      put :promote, params: { id: @lead.id, access: "Lead", account: { name: "Hello", access: "Lead", user_id: current_user.id }, opportunity: { name: "World", access: "Lead", user_id: current_user.id } }, xhr: true
       expect(@account.access).to eq("Shared")
       expect(@account.permissions.map(&:user_id).sort).to eq([7, 8])
       expect(@account.permissions.map(&:asset_id)).to eq([@account.id, @account.id])
@@ -757,14 +757,14 @@ describe LeadsController do
       @campaign = FactoryGirl.create(:campaign)
       @lead = FactoryGirl.create(:lead, user: current_user, campaign: @campaign)
 
-      xhr :put, :promote, id: @lead.id, account: { name: "Hello" }, opportunity: { name: "Hello", campaign_id: @campaign.id }
+      put :promote, params: { id: @lead.id, account: { name: "Hello" }, opportunity: { name: "Hello", campaign_id: @campaign.id } }, xhr: true
       expect(assigns[:opportunity].campaign).to eq(@campaign)
     end
 
     it "should assign lead's source to the newly created opportunity" do
       @lead = FactoryGirl.create(:lead, user: current_user, source: "cold_call")
 
-      xhr :put, :promote, id: @lead.id, account: { name: "Hello" }, opportunity: { name: "Hello", source: @lead.source }
+      put :promote, params: { id: @lead.id, account: { name: "Hello" }, opportunity: { name: "Hello", source: @lead.source } }, xhr: true
       expect(assigns[:opportunity].source).to eq(@lead.source)
     end
 
@@ -772,7 +772,7 @@ describe LeadsController do
       @lead = FactoryGirl.create(:lead)
       request.env["HTTP_REFERER"] = "http://localhost/leads"
 
-      xhr :put, :promote, id: @lead.id, account: { name: "Hello" }, opportunity: {}
+      put :promote, params: { id: @lead.id, account: { name: "Hello" }, opportunity: {} }, xhr: true
       expect(assigns[:lead_status_total]).not_to be_nil
       expect(assigns[:lead_status_total]).to be_an_instance_of(HashWithIndifferentAccess)
     end
@@ -782,7 +782,7 @@ describe LeadsController do
       @lead = FactoryGirl.create(:lead, campaign: @campaign)
       request.env["HTTP_REFERER"] = "http://localhost/campaigns/#{@campaign.id}"
 
-      xhr :put, :promote, id: @lead.id, account: { name: "Hello" }, opportunity: {}
+      put :promote, params: { id: @lead.id, account: { name: "Hello" }, opportunity: {} }, xhr: true
       expect(assigns[:campaign]).to eq(@campaign)
     end
 
@@ -792,7 +792,7 @@ describe LeadsController do
       @contact = FactoryGirl.build(:contact, first_name: nil) # make it fail
       allow(Contact).to receive(:new).and_return(@contact)
 
-      xhr :put, :promote, id: 42, account: { id: 123 }
+      put :promote, params: { id: 42, account: { id: 123 } }, xhr: true
       expect(@lead.reload.status).to eq("new")
       expect(response).to render_template("leads/promote")
     end
@@ -802,7 +802,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, user: current_user)
         @lead.destroy
 
-        xhr :put, :promote, id: @lead.id
+        put :promote, params: { id: @lead.id }, xhr: true
         expect(flash[:warning]).not_to eq(nil)
         expect(response.body).to eq("window.location.reload();")
       end
@@ -810,7 +810,7 @@ describe LeadsController do
       it "should reload current page with the flash message if the lead is protected" do
         @private = FactoryGirl.create(:lead, user: FactoryGirl.create(:user), access: "Private")
 
-        xhr :put, :promote, id: @private.id
+        put :promote, params: { id: @private.id }, xhr: true
         expect(flash[:warning]).not_to eq(nil)
         expect(response.body).to eq("window.location.reload();")
       end
@@ -827,7 +827,7 @@ describe LeadsController do
 
     describe "AJAX request" do
       it "should reject the requested lead and render [reject] template" do
-        xhr :put, :reject, id: @lead.id
+        put :reject, params: { id: @lead.id }, xhr: true
 
         expect(assigns[:lead]).to eq(@lead.reload)
         expect(@lead.status).to eq("rejected")
@@ -836,7 +836,7 @@ describe LeadsController do
 
       it "should get the data for leads sidebar when called from leads index" do
         request.env["HTTP_REFERER"] = "http://localhost/leads"
-        xhr :put, :reject, id: @lead.id
+        put :reject, params: { id: @lead.id }, xhr: true
         expect(assigns[:lead_status_total]).not_to be_nil
         expect(assigns[:lead_status_total]).to be_an_instance_of(HashWithIndifferentAccess)
       end
@@ -846,7 +846,7 @@ describe LeadsController do
         @lead = FactoryGirl.create(:lead, campaign: @campaign)
 
         request.env["HTTP_REFERER"] = "http://localhost/campaigns/#{@campaign.id}"
-        xhr :put, :reject, id: @lead.id
+        put :reject, params: { id: @lead.id }, xhr: true
         expect(assigns[:campaign]).to eq(@campaign)
       end
 
@@ -855,7 +855,7 @@ describe LeadsController do
           @lead = FactoryGirl.create(:lead, user: current_user)
           @lead.destroy
 
-          xhr :put, :reject, id: @lead.id
+          put :reject, params: { id: @lead.id }, xhr: true
           expect(flash[:warning]).not_to eq(nil)
           expect(response.body).to eq("window.location.reload();")
         end
@@ -863,7 +863,7 @@ describe LeadsController do
         it "should reload current page with the flash message if the lead is protected" do
           @private = FactoryGirl.create(:lead, user: FactoryGirl.create(:user), access: "Private")
 
-          xhr :put, :reject, id: @private.id
+          put :reject, params: { id: @private.id }, xhr: true
           expect(flash[:warning]).not_to eq(nil)
           expect(response.body).to eq("window.location.reload();")
         end
@@ -872,7 +872,7 @@ describe LeadsController do
 
     describe "HTML request" do
       it "should redirect to Leads index when a lead gets rejected from its landing page" do
-        put :reject, id: @lead.id
+        put :reject, params: { id: @lead.id }
 
         expect(assigns[:lead]).to eq(@lead.reload)
         expect(@lead.status).to eq("rejected")
@@ -885,7 +885,7 @@ describe LeadsController do
           @lead = FactoryGirl.create(:lead, user: current_user)
           @lead.destroy
 
-          put :reject, id: @lead.id
+          put :reject, params: { id: @lead.id }
           expect(flash[:warning]).not_to eq(nil)
           expect(response).to redirect_to(leads_path)
         end
@@ -893,7 +893,7 @@ describe LeadsController do
         it "should redirect to lead index if the lead is protected" do
           @private = FactoryGirl.create(:lead, user: FactoryGirl.create(:user), access: "Private")
 
-          put :reject, id: @private.id
+          put :reject, params: { id: @private.id }
           expect(flash[:warning]).not_to eq(nil)
           expect(response).to redirect_to(leads_path)
         end
@@ -954,7 +954,7 @@ describe LeadsController do
   #----------------------------------------------------------------------------
   describe "responding to GET redraw" do
     it "should save user selected lead preference" do
-      xhr :get, :redraw, per_page: 42, view: "long", sort_by: "first_name", naming: "after"
+      get :redraw, params: { per_page: 42, view: "long", sort_by: "first_name", naming: "after" }, xhr: true
       expect(current_user.preference[:leads_per_page]).to eq("42")
       expect(current_user.preference[:leads_index_view]).to eq("long")
       expect(current_user.preference[:leads_sort_by]).to eq("leads.first_name ASC")
@@ -962,13 +962,13 @@ describe LeadsController do
     end
 
     it "should set similar options for Contacts" do
-      xhr :get, :redraw, sort_by: "first_name", naming: "after"
+      get :redraw, params: { sort_by: "first_name", naming: "after" }, xhr: true
       expect(current_user.pref[:contacts_sort_by]).to eq("contacts.first_name ASC")
       expect(current_user.pref[:contacts_naming]).to eq("after")
     end
 
     it "should reset current page to 1" do
-      xhr :get, :redraw, per_page: 42, view: "long", sort_by: "first_name", naming: "after"
+      get :redraw, params: { per_page: 42, view: "long", sort_by: "first_name", naming: "after" }, xhr: true
       expect(session[:leads_current_page]).to eq(1)
     end
 
@@ -978,7 +978,7 @@ describe LeadsController do
         FactoryGirl.create(:lead, first_name: "Bobby", user: current_user)
       ]
 
-      xhr :get, :redraw, per_page: 1, sort_by: "first_name"
+      get :redraw, params: { per_page: 1, sort_by: "first_name" }, xhr: true
       expect(assigns(:leads)).to eq([@leads.first])
       expect(response).to render_template("leads/index")
     end
@@ -991,7 +991,7 @@ describe LeadsController do
       session[:leads_filter] = "contacted,rejected"
 
       @leads = [FactoryGirl.create(:lead, user: current_user, status: "new")]
-      xhr :post, :filter, status: "new"
+      post :filter, params: { status: "new" }, xhr: true
       expect(assigns[:leads]).to eq(@leads)
       expect(response).to be_a_success
       expect(response).to render_template("leads/index")
@@ -999,7 +999,7 @@ describe LeadsController do
 
     it "should reset current page to 1" do
       @leads = []
-      xhr :post, :filter, status: "new"
+      post :filter, params: { status: "new" }, xhr: true
 
       expect(session[:leads_current_page]).to eq(1)
     end
