@@ -36,7 +36,7 @@ describe 'FatFreeCRM::Fields' do
 
   describe "field_groups" do
     it "should call FieldGroup" do
-      expect(ActiveRecord::Base.connection).to receive(:table_exists?).with('field_groups').and_return(true)
+      expect(ActiveRecord::Base.connection).to receive(:data_source_exists?).with('field_groups').and_return(true)
       dummy_scope = double
       expect(dummy_scope).to receive(:order).with(:position)
       expect(FieldGroup).to receive(:where).and_return(dummy_scope)
@@ -44,7 +44,7 @@ describe 'FatFreeCRM::Fields' do
     end
 
     it "should not call FieldGroup if table doesn't exist (migrations not yet run)" do
-      expect(ActiveRecord::Base.connection).to receive(:table_exists?).with('field_groups').and_return(false)
+      expect(ActiveRecord::Base.connection).to receive(:data_source_exists?).with('field_groups').and_return(false)
       expect(Foo.new.field_groups).to eq([])
     end
   end
