@@ -10,13 +10,13 @@ shared_examples "auto complete" do
   end
 
   it "should do the search and find records that match autocomplete query" do
-    get :auto_complete, auto_complete_query: @query
+    get :auto_complete, params: { auto_complete_query: @query }
     expect(assigns[:query]).to eq(@query)
     expect(assigns[:auto_complete]).to eq(@auto_complete_matches) # Each controller must define it.
   end
 
   it "should save current autocomplete controller in a session" do
-    get :auto_complete, auto_complete_query: @query
+    get :auto_complete, params: { auto_complete_query: @query }
 
     # We don't save Admin/Users autocomplete controller in a session since Users are not
     # exposed through the Jumpbox.
@@ -26,14 +26,14 @@ shared_examples "auto complete" do
   end
 
   it "should render application/_auto_complete template" do
-    post :auto_complete, auto_complete_query: @query
+    post :auto_complete, params: { auto_complete_query: @query }
     expect(response).to render_template("application/_auto_complete")
   end
 end
 
 shared_examples "attach" do
   it "should attach existing asset to the parent asset of different type" do
-    xhr :put, :attach, id: @model.id, assets: @attachment.class.name.tableize, asset_id: @attachment.id
+    put :attach, params: { id: @model.id, assets: @attachment.class.name.tableize, asset_id: @attachment.id }, xhr: true
     expect(@model.send(@attachment.class.name.tableize)).to include(@attachment)
     expect(assigns[:attachment]).to eq(@attachment)
     expect(assigns[:attached]).to eq([@attachment])
@@ -53,7 +53,7 @@ shared_examples "attach" do
       @model.send(@attachment.class.name.tableize) << @attachment
     end
 
-    xhr :put, :attach, id: @model.id, assets: @attachment.class.name.tableize, asset_id: @attachment.id
+    put :attach, params: { id: @model.id, assets: @attachment.class.name.tableize, asset_id: @attachment.id }, xhr: true
     expect(assigns[:attached]).to eq(nil)
     expect(response).to render_template("entities/attach")
   end
@@ -61,14 +61,14 @@ shared_examples "attach" do
   it "should display flash warning when the model is no longer available" do
     @model.destroy
 
-    xhr :put, :attach, id: @model.id, assets: @attachment.class.name.tableize, asset_id: @attachment.id
+    put :attach, params: { id: @model.id, assets: @attachment.class.name.tableize, asset_id: @attachment.id }, xhr: true
     expect(flash[:warning]).not_to eq(nil)
     expect(response.body).to eq("window.location.reload();")
   end
   it "should display flash warning when the attachment is no longer available" do
     @attachment.destroy
 
-    xhr :put, :attach, id: @model.id, assets: @attachment.class.name.tableize, asset_id: @attachment.id
+    put :attach, params: { id: @model.id, assets: @attachment.class.name.tableize, asset_id: @attachment.id }, xhr: true
     expect(flash[:warning]).not_to eq(nil)
     expect(response.body).to eq("window.location.reload();")
   end
@@ -76,7 +76,7 @@ end
 
 shared_examples "discard" do
   it "should discard the attachment without deleting it" do
-    xhr :post, :discard, id: @model.id, attachment: @attachment.class.name, attachment_id: @attachment.id
+    post :discard, params: { id: @model.id, attachment: @attachment.class.name, attachment_id: @attachment.id }, xhr: true
     expect(assigns[:attachment]).to eq(@attachment.reload)                     # The attachment should still exist.
     expect(@model.reload.send(@attachment.class.name.tableize.to_s)).to eq([]) # But no longer associated with the model.
     expect(assigns[:account]).to eq(@model) if @model.is_a?(Account)
@@ -88,7 +88,7 @@ shared_examples "discard" do
   it "should display flash warning when the model is no longer available" do
     @model.destroy
 
-    xhr :post, :discard, id: @model.id, attachment: @attachment.class.name, attachment_id: @attachment.id
+    post :discard, params: { id: @model.id, attachment: @attachment.class.name, attachment_id: @attachment.id }, xhr: true
     expect(flash[:warning]).not_to eq(nil)
     expect(response.body).to eq("window.location.reload();")
   end
@@ -96,7 +96,7 @@ shared_examples "discard" do
   it "should display flash warning when the attachment is no longer available" do
     @attachment.destroy
 
-    xhr :post, :discard, id: @model.id, attachment: @attachment.class.name, attachment_id: @attachment.id
+    post :discard, params: { id: @model.id, attachment: @attachment.class.name, attachment_id: @attachment.id }, xhr: true
     expect(flash[:warning]).not_to eq(nil)
     expect(response.body).to eq("window.location.reload();")
   end
