@@ -12,7 +12,7 @@ describe "/leads/index" do
     view.lookup_context.prefixes << 'entities'
     assign :per_page, Lead.per_page
     assign :sort_by,  Lead.sort_by
-    assign :ransack_search, Lead.search
+    assign :ransack_search, Lead.ransack
     login_and_assign
   end
 
